File size optimization
In working with this Framework I have found that the framework could use some optimizations.
Sadly this change does require the short tags functionality enabled of PHP, but I figured it'd be worth posting nonetheless.

In this change, I have managed to reduce the filesize of the framework by 12%, effectively decreasing the parsing time required. Most developers will see a notable performance increase when benchmarking this >1.000.000 times.